refact: enums moved in to a new package
