+++ src/main/java/com/healingsys/entities/enums/Role.java
@@ -1,4 +1,4 @@
-package com.healingsys.entities;
+package com.healingsys.entities.enums;
 
 public enum Role {
     PATIENT,
